Rename Kql.* to kql.*
